Make FinerStruct behave like a value object
Structs and OpenStructs are used to represent value objects, where they
are defined by their values. If we had two objects with the same values,
that would mean they are equal. This was not the case:

```ruby
Q = FinerStruct::Immutable(:value)
a = Q.new(value:1)
=> #<Q:0x007fa1123d44c8 @value=1>
b = Q.new(value:1)
=> #<Q:0x007fa1122ed938 @value=1>
a == b
=> false
```

This PR overides the existing equality (#==) method to compare its
values to another object's values. If they are all equal it will return
true.